+++ Core/Math/Vector.cs
@@ -3,7 +3,7 @@ using System;
 namespace FIVES
 {
     public struct Vector {
-        public float x, y, z;
+        public double x, y, z;
     }
 }
