+++ Plugins/Motion/RotVelocity.cs
@@ -5,7 +5,7 @@ namespace MotionPlugin
 {
     public struct RotVelocity {
         public Vector axis;
-        public float rotSpeed;
+        public double rotSpeed;
     }
 }
